Use config logo for local icon
