+++ app/soapbox/hooks/useTheme.ts
@@ -17,4 +17,6 @@ const useTheme = (): Theme => {
   return darkMode ? 'dark' : 'light';
 };
 
+
+
 export { useTheme };
